+++ aqua/diagnostics/base/diagnostic.py
@@ -291,7 +291,7 @@ class Diagnostic():
 
         return region, lon_limits, lat_limits
 
-    def _select_region(self, data: xr.Dataset, region: str = None, diagnostic: str = None, drop: bool = True, **kwargs):
+    def select_region(self, data: xr.Dataset, region: str = None, diagnostic: str = None, drop: bool = True, **kwargs):
         """
         Select a geographic region from the dataset. Used when selection is not on the self.data attribute.