+++ aqua/diagnostics/ocean_trends/trends.py
@@ -91,7 +91,7 @@ class Trends(Diagnostic):
         # If a region is specified, apply area selection to self.data
         if region:
             self.logger.info(f"Selecting region: {region}.")
-            res_dict = super()._select_region(
+            res_dict = super().select_region(
                 data=data, region=region, diagnostic="ocean3d", drop=True
             )
             lat_limits = res_dict["lat_limits"]